+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ UNRELEASED
 Requires libfranka >= 0.20.4 and franka_description >= 2.4.0 requires ROS 2 Jazzy
 
 * feat: integration_launch_testing: added smoke tests for the example controllers
+* feat: franka_mobile_sensors: add visualization of sensors in rviz
 * fix: gripper_example_controller also works without namespace
 * fix: corrected logs in franka_hardware 
 * fix: gravity_compensation_example_controller, move_to_start_example_controller, joint_impedance_example_controller work with parametrized robot_type